Dev 18-07-24-02
Figured out why the program was unable to write to appdata. IT WAS BECUASE OF MICROSOFT, CAUSE OF COURSE IT WAS! Never use microsoft store. I'm tempted to reinstall minecraft now and see if there's a way to force a non-windows store minecraft launcher instead, cause everyone knows the windows store sucks. THIS IS WHY!!!